+++ src/textures/mod.rs
@@ -1,7 +1,7 @@
 use crate::vec3::{Color, Point3};
 
 pub trait Texture: Send + Sync {
-    fn value(&self, uv: (f64, f64), p: Point3) -> Color;
+    fn value(&self, uv: (f32, f32), p: Point3) -> Color;
 }
 
 pub mod checker_texture;
